update jenkins
starting to work on both branches , dev and staging , assuming the dev branch role ends when a docker image of the app is build and pushed to docker hub "Continous integration". as for staging , supposedly the QA team will connect to the app through the container (image source from dockerhub ) ,when testing fails , staging team will raise a ticket and start again from dev , if a success , we discuss that later , still don't have the right assumptions to make. for now the deployment will be on master (ubuntu 20.04) , untill i link my vm so i can have a slave agent .